+++ doc/classes/@GlobalScope.xml
@@ -44,6 +44,8 @@
 		<member name="JSON" type="JSON" setter="" getter="">
 			[JSON] singleton
 		</member>
+		<member name="JavaScript" type="JavaScript" setter="" getter="">
+		</member>
 		<member name="Marshalls" type="Reference" setter="" getter="">
 			[Marshalls] singleton
 		</member>
